fix(EIF): use float-precision hyperplane weights for Spark 4.x Avro compatibility
  Spark 4.x's Avro encoder silently demotes Array[Double] elements to
  float (32-bit) precision during serialization, while scalar Double
  fields survive intact. This caused all five EIF model write/read tests
  to fail on Spark 4.0.1 and 4.1.1, with weight mismatches at ~1e-8
  (the exact double→float→double precision boundary).

  The fix changes SplitHyperplane weights from Array[Double] to
  Array[Float]. This is the correct design from first principles:
  - Features are already Array[Float] (DataPoint.features)
  - Weights define the hyperplane *direction* (analogous to the feature
    index in standard IF, which is just an Int)
  - The offset defines *where* to split and remains Double (analogous to
    splitValue in standard IF, which is Double for the same reason)
  - The dot product is accumulated in Double regardless of operand type
  - The split comparison (dot < offset) is always Double vs Double

  Weights are converted to float after normalization but before computing
  the offset, so training and scoring are consistent. Benchmarks confirm
  the change is invisible after rounding: only one value across all 13
  datasets changed (breastw ExtendedIF_max AUPRC: 0.9568 → 0.9569,
  well within the ±0.0015 error bar).

  Production code:
  - ExtendedUtils.scala: SplitHyperplane.weights Array[Double] → Array[Float]
  - ExtendedIsolationTree.scala: normalize to float before offset computation
  - ExtendedIsolationForestModelReadWrite.scala: ExtendedNodeData.weights
    and NullWeights updated to float

  Test code:
  - ExtendedIsolationTreeTest.scala: float literals, L2 norm tolerance
    widened from 1e-10 to 1e-6 (appropriate for float precision)
  - ExtendedIsolationForestModelWriteReadTest.scala: float literals,
    added disabled regenerateGoldenExtendedModel helper
  - Regenerated golden model and expected tree structure

  README:
  - Updated breastw ExtendedIF_max AUPRC from 0.9568 to 0.9569

  Verified all 67 tests pass on Spark 3.5.5, 4.0.1, and 4.1.1.

+++ isolation-forest/src/main/scala/com/linkedin/relevance/isolationforest/extended/ExtendedIsolationTree.scala
@@ -183,16 +183,21 @@ private[isolationforest] object ExtendedIsolationTree extends Logging {
         if (normValue == 0) {
           ExtendedExternalNode(numInstances)
         } else {
-          // Normalize the vector
-          val normalizedWeights = new Array[Double](rawWeights.length)
+          // Normalize the vector and convert to float. Float precision is sufficient
+          // for random hyperplane directions (features are already float), and the
+          // double-precision offset handles the "where to split" precision — mirroring
+          // how standard IF uses float features with a double splitValue.
+          val normalizedWeights = new Array[Float](rawWeights.length)
           i = 0
           while (i < rawWeights.length) {
-            normalizedWeights(i) = rawWeights(i) / normValue
+            normalizedWeights(i) = (rawWeights(i) / normValue).toFloat
             i += 1
           }
 
           // Sample intercept point p from the per-coordinate ranges.
           // Only coordinates with non-zero weights in the hyperplane matter.
+          // The offset is computed with the float weights so that training and
+          // scoring are consistent.
           var splitOffset = 0.0
           var k = 0
           while (k < sparseIndices.length) {
